+++ src/main/java/com/ae2powertools/features/monitor/dependent/IStorageMonitorHost.java
@@ -46,6 +46,10 @@ public interface IStorageMonitorHost {
 
     void setMatchMode(MatchMode mode);
 
+    boolean isHysteresisEnabled();
+
+    void setHysteresisEnabled(boolean hysteresisEnabled);
+
     // --- Condition ---
 
     /**